backend/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis (Job Queue & Caching)
redis==5.0.1
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
psutil==5.9.8

# WebSocket for real-time updates
websockets==12.0

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Existing automation dependencies
openpyxl==3.1.2
selenium==4.16.0
webdriver-manager==4.0.1
pandas==2.1.4
colorama==0.4.6

# Email notifications
emails==0.6

# Utilities
httpx==0.26.0
pydantic-extra-types==2.4.1

# Funda automation (DrissionPage + Google Sheets)
DrissionPage==4.1.1.2
gspread==6.2.1
google-auth==2.49.0
